Split port parsing logic from pod specific logic. Move pods logic into inbound
