+++ policy-controller/k8s/index/src/inbound.rs
@@ -3,6 +3,7 @@ mod http_route;
 pub mod index;
 mod meshtls_authentication;
 mod network_authentication;
+mod pod;
 mod server;
 pub mod server_authorization;
